backend/requirements.txt
Flask==3.0.0
Flask-CORS==6.0.1
python-dotenv==1.2.1
openai==2.8.1
firecrawl-py==4.10.1
requests==2.32.5
gunicorn==23.0.0
pydantic==2.12.5
gevent